Fixed the issue with the addSignal function
Turns out that it wasn't an issue with the function, but the naming of the World class, module, and instance. The function addSignal belongs the the World module, as does the instance of World, pointed at by a lowercase w world. It gets a bit confusing and perhaps should be changed. Nonetheless, it works now.