Fixed the issue with the Camera...Kind of
So I was able to slightly remedy the issue we were having in class, in that you can now change the Camera's position using actual reactive programming (position = p3(x,y,z)). However, two Cameras still exist in the world. It has something to do with the import that looks like:

 from PandaSRC.Camera import camera

 For some reason this import creates an instance of Camera, and it currently exists in two places (Bezier and PandaEngine) hence the two instances of Camera. Removing the PandaEngine, which we are going to do anyway, should fix this dual camera 'issue' (if you can still call it that) without any problems.