Changed PandaEngine and fixed a resetWorld error
I altered the removed the PandaEngine file from the repo completely, and replaced it with PandaHeartbeat. The idea being that any needed engine functionality is exported from frp, rather than copied into Reactive Panda. There is something strange going on in this file if I understand it correctly; put bluntly, it shouldn't work, but it does. Why is that?

Also, the resetWorld function was missing the ability to receive  a function to continue on to as an argument. This was my error from when I re-did the World file earlier. A simple fix., simply add the function as an argument in PandaWorld.