Found out where the objects are being added & edited demos
So I found out that all objects created by the Proxy init are automatically added to the worldObjects list, thus answering there question of 'Where is the stuff that theoretically went in PandaEngine happening?'.  I also deleted a few demos that were not so well showcasing the engine, and edited two that I believe do a good job of showcasing the proper features. More demo work still needs to be done.